Improved Functionallity and User Experience
Output file now displays 1 decimal point accuracy; CouseFile.txt and NameFile.txt are placed in the default folder (same path as project folder) which the program will use if no input file is specified. New priorities are error handling for incorrect data, displaying selected folder in the UI, and conveying errors to the user (ex. if output file with the same name exists)